routers refactoring, minio storage proxy
