+++ gcs-minio-proxy/Dockerfile
@@ -1,0 +1,2 @@
+FROM minio/minio
+ENTRYPOINT ["sh", "-c", "exec minio gateway gcs $GCLOUD_PROJECT_ID"]
